Remove redundant validation on trust corporation fields
Given that name and number are both fields, having a NotBlank
constraint on these fields means they both have to have at least
one character in them (so they are not blank). It's therefore
redundant also having a minimum length 1 constraint, as that
means the same thing as not blank in this context.

This allows us to remove an awkward work-around in Make which has
to ignore duplicate error messages for the same problem (an empty
string value in the name or number of a trust corporation).

(NB if anyone has a preference, we could remove the NotBlank constraint
and keep the min length 1 constraint instead.)